+++ CHANGELOG.md
@@ -4,8 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.0 - 2021-07-03
+### Added
+- Method attribute analysis.
+- Ability to register multiple classes at once.
+
 ## 1.0.0 - 2021-02-23
 ### Added
 - Initial implementation of the package.
 
-[1.0.0 > Unreleased](https://github.com/grizz-it/metadata/compare/1.0.0...HEAD)
+[1.1.0 > Unreleased](https://github.com/grizz-it/metadata/compare/1.1.0...HEAD)
+[1.0.0 > 1.1.0](https://github.com/grizz-it/metadata/compare/1.0.0...1.1.0)